refactor(genetics): split behavioral traits module into focused units
Implements IMPROVEMENT_PROPOSALS.md 2.2 (fourth god-file split):
core/genetics/behavioral.py drops from 830 to 270 lines. Trait specs
stay declarative in behavioral.py; the logic moves to named,
individually testable modules:

- behavioral_inheritance.py: inherit/recombine trait paths (verbatim
  bodies, specs passed as a parameter), composable-behavior and
  poker-strategy inheritance helpers
- mate_preferences.py: preference specs, normalization, inheritance
- policy_inheritance.py: code-policy inheritance, mutation, validation

behavioral.py re-exports every previous name (including underscore
aliases tests import), so the public surface is unchanged. Note: the
backlog suggested a BehavioralTraitCodec, but (de)serialization already
lives in genome_codec.py - the policy module took its place.

Verified: full fast gate 1699 passed (matches baseline exactly); golden
replay passes; tank/survival_5k champion reproduces exactly (RNG call
order preserved); mypy clean (333 files).
